+++ Source/BuildPackages/Templates/Web_contains.pas
@@ -1,0 +1,2 @@
+contains
+  NovusWebUtils in '..\..\Core\Utilities\NovusWebUtils.pas';