+++ Source/BuildPackages/Templates/Web_requires.pas
@@ -1,0 +1,4 @@
+requires
+  rtl,
+  NovusCodeLibrary_Core,
+  NovusCodeLibrary_Env;