+++ Source/BuildPackages/Templates/cURL_contains.pas
@@ -1,0 +1,2 @@
+contains
+  NovuscURLUtils in '..\..\Core\Utilities\cURL\NovuscURLUtils.pas';