+++ Source/BuildPackages/Templates/cURL_requires.pas
@@ -1,0 +1,5 @@
+requires
+  rtl,
+  NovusCodeLibrary_Core,
+  NovusCodeLibrary_WebUtils,
+  NovusCodeLibrary_Env;